Fix FavoritesScreen
