+++ pubspec.yaml
@@ -9,38 +9,51 @@ environment:
   sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
+  
+  flutter:
+    sdk: flutter
+
+  # State management
+  flutter_bloc: ^8.0.0
+  flutter_mobx: ^2.0.2
+  flutter_redux: ^0.8.2
+  mobx: ^2.0.5
+  mwwm: ^2.0.0
+  redux: ^5.0.0
+  redux_thunk: ^0.4.0
+
+  # Network
   dio: ^4.0.0
   easy_localization: ^3.0.0
   equatable: ^2.0.3
   file_picker: ^4.2.1
-  flutter:
-    sdk: flutter
-  flutter_bloc: ^8.0.0
-  flutter_mobx: ^2.0.2
-  flutter_redux: ^0.8.2
+
   flutter_svg: ^0.22.0
   flutter_switch: ^0.3.2
   image_picker: ^0.8.4+4
-  cached_network_image: ^3.2.0
+  
   intl: ^0.17.0
   mime_type: ^1.0.0
-  mobx: ^2.0.5
-  mwwm: ^2.0.0
+  
   open_file: ^3.2.1
   photo_view: ^0.13.0
   provider: ^6.0.1
-  redux: ^5.0.0
-  redux_thunk: ^0.4.0
+  
   relation: ^3.1.0
   rxdart: ^0.27.3
+
+  # cache images
+  cached_network_image: ^3.2.0
   
 
 dev_dependencies:
-  dart_code_metrics: ^4.6.0
-  build_runner: ^2.1.4
+  
   flutter_test:
     sdk: flutter
+    
   mobx_codegen: ^2.0.4
+  dart_code_metrics: ^4.6.0
+  build_runner: ^2.1.4
 
 
 flutter:
